refactor(model): split downloadModels.py into registry.py + download.py
Separate the two concerns in the 1347-line file:
  src/model/registry.py  -- weightsBaseDir, weightsDir, modelsList, modelsMap
                            (pure static data, no I/O, safe to import anywhere)
  src/model/download.py  -- URL constants, downloadAndLog, downloadModels,
                            resolveWeightPath (network + filesystem I/O)

src/model/downloadModels.py is kept as a combined re-export shim so all
existing callers continue working without change.
